feat(TurboRangoSolution) - Alterando latitude para double

+++ src/TurboRangoSolution/TurboRango.Dominio/Localizacao.cs
@@ -8,7 +8,7 @@ namespace TurboRango.Dominio
     internal class Localizacao
     {
         internal string bairro;
-        internal int latitude;
+        internal double latitude;
         internal string logradouro;
         internal int longitude;
     }
